refactor(ci): fix the CI layer inversion
The GitHub Forge::fetch_checks arm called UP into commands::ci, so the
lowest abstraction (raw GitHub check-run HTTP) lived in the top command
file and the dependency ran application -> forge -> commands.

- Move GitHub check-run fetching into the adapter (src/github/checks.rs,
  impl GitHubClient) so the Forge arm calls its own inherent method,
  symmetric with the GitLab/Gitea arms.
- Move the provider-neutral CheckRunInfo transforms (dedup, commit-status
  normalization) into src/ci/normalize.rs so fetch and render share one
  copy.

Dependency gradient now points down. No user-visible behavior change; the
commands/ci.rs file split was deliberately deferred (test module too
entangled to split safely).
